automation/requirements.txt
fastapi
uvicorn
twilio
openai
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
python-multipart
